+++ R/tbl_lazy_api.R
@@ -1,6 +1,6 @@
 #' @importFrom vctrs vec_ptype_abbr
 #' @export
-vec_ptype_abbr.collect <- function(x) {
+vec_ptype_abbr.collect <- function(x, ...) {
   "collect"
 }